fix: Rename logging.py, correct log output format
